feat(server): move resolvers to separate file and folders
