+++ server/src/apollo-server.ts
@@ -7,6 +7,7 @@ import { GraphQLFileLoader } from '@graphql-tools/graphql-file-loader';
 import { loadSchemaSync } from '@graphql-tools/load';
 import { addResolversToSchema } from '@graphql-tools/schema';
 import { GRAPHQL_SCHEMA_PATH } from './constants';
+import resolvers from './resolvers';
 
 const SCHEMA = loadSchemaSync(GRAPHQL_SCHEMA_PATH, {
   loaders: [new GraphQLFileLoader()],
@@ -17,25 +18,6 @@ export async function createApolloServer(
   httpServer: Server,
   app: express.Application
 ): Promise<ApolloServer<ExpressContext>> {
-  const resolvers = {
-    Query: {
-      currentUser: () => {
-        return {
-          id: '123',
-          name: 'John Doe',
-          handle: 'johndoe',
-          coverUrl: '',
-          avatarUrl: '',
-          createdAt: '',
-          updatedAt: '',
-        };
-      },
-      suggestions: () => {
-        return [];
-      },
-    },
-  };
-
   const server = new ApolloServer({
     schema: addResolversToSchema({ schema: SCHEMA, resolvers }),
     context: () => ({ db }),
