+++ server/src/resolvers.ts
@@ -1,0 +1,7 @@
+import queryTwitterResolvers from './resolvers/Query';
+
+const resolvers = {
+  Query: queryTwitterResolvers,
+};
+
+export default resolvers;
