+++ server/src/resolvers/Query.ts
@@ -1,0 +1,18 @@
+const queryTwitterResolvers = {
+  currentUser: () => {
+    return {
+      id: '123',
+      name: 'John Doe',
+      handle: 'johndoe',
+      coverUrl: '',
+      avatarUrl: '',
+      createdAt: '',
+      updatedAt: '',
+    };
+  },
+  suggestions: () => {
+    return [];
+  },
+};
+
+export default queryTwitterResolvers;
